chore: remove ImageFileType
